+++ packages/react-mui-authn/rollup.config.mjs
@@ -38,7 +38,7 @@ export default [
     external: Object.keys(pkg.dependencies),
   },
   {
-    input: "dist/esm/index.d.ts",
+    input: "dist/esm/types/index.d.ts",
     output: [{ file: "dist/index.d.ts", format: "esm" }],
     external: [/\.css$/],
     plugins: [dts()],