+++ packages/react-mui-shared/rollup.config.mjs
@@ -1,4 +1,3 @@
-// /rollup.config.js
 import resolve from "@rollup/plugin-node-resolve";
 import external from "rollup-plugin-peer-deps-external";
 import terser from "@rollup/plugin-terser";
@@ -7,22 +6,23 @@ import typescript from "@rollup/plugin-typescript";
 import commonjs from "@rollup/plugin-commonjs";
 import replace from "@rollup/plugin-replace";
 import nodePolyfills from "rollup-plugin-polyfill-node";
-import dts from "rollup-plugin-dts";
+import { dts } from "rollup-plugin-dts";
 
-import pkg from "./package.json";
+import pkg from "./package.json" assert { type: "json" };
 
 export default [
   {
     input: "src/index.ts",
     output: [
       {
-        file: pkg.main,
+        dir: "dist/cjs",
         format: "cjs",
         sourcemap: true,
+        interop: "auto",
         name: "react-lib",
       },
       {
-        file: pkg.module,
+        dir: "dist/esm",
         format: "esm",
         sourcemap: true,
       },
@@ -41,10 +41,9 @@ export default [
       }),
     ],
     external: Object.keys(pkg.peerDependencies || {}),
-    inlineDynamicImports: true,
   },
   {
-    input: "dist/esm/index.d.ts",
+    input: "src/index.ts",
     output: [{ file: "dist/index.d.ts", format: "esm" }],
     external: [/\.css$/],
     plugins: [dts()],