+++ packages/react-mui-audit-log-viewer/CHANGELOG.md
@@ -5,6 +5,48 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.70] - 2024-02-28
+
+### Changed
+
+- Update to work with latest download endpoint
+
+## [0.0.69] - 2024-02-26
+
+### Changed
+
+- Update to @pangeacyber/react-mui-share@0.0.54
+
+## [0.0.67] - 2024-02-23
+
+### Changed
+
+- Switch download to json
+
+## [0.0.65] - 2024-02-22
+
+### Changed
+
+- Only support download if there is more than one result
+
+## [0.0.64] - 2024-02-22
+
+### Changed
+
+- Update to include optional onDownload callback
+
+## [0.0.63] - 2024-02-22
+
+### Changed
+
+- Update to @pangeacyber/react-mui-shared@0.0.50
+
+## [0.0.62] - 2024-02-15
+
+### Fixed
+
+- Color to lighten/darken based on theme mode
+
 ## [0.0.60] - 2024-01-24
 
 ### Fixed