+++ packages/react-mui-store-file-viewer/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.33] - 2024-02-26
+
+### Fixed
+
+- Update to @pangeacyber/react-mui-share@0.0.54
+
+## [0.0.32] - 2024-02-22
+
+### Fixed
+
+- Update to @pangeacyber/react-mui-share@0.0.52
+
+## [0.0.31] - 2024-02-15
+
+### Fixed
+
+- Color to lighten/darken based on theme mode
+
 ## [0.0.30] - 2024-02-14
 
 ### Added